Refactor: rename fille name & small changes in scss
